update SqlManager.java & add SqlUtil.java
